seeder fixed

+++ app/Config/Database.php
@@ -126,7 +126,7 @@ class Database extends Config
             'compress'     => false,
             'strictOn'     => false,
             'failover'     => [],
-            'port'         => env('database.default.port', 3306),
+            'port'         => (int) env('database.default.port', 3306),
             'numberNative' => false,
         ];
 
@@ -149,7 +149,7 @@ class Database extends Config
                 'compress'    => false,
                 'strictOn'    => false,
                 'failover'    => [],
-                'port'        => env('database.tests.port', 3306),
+                'port'        => (int) env('database.tests.port', 3306),
                 'foreignKeys' => true,
                 'busyTimeout' => 1000,
             ];
